Comleted: back end testing, Z-score calculation, and threshold calculation.
the highest return that was possible with the code at this moment is %1.65.

while still in the green there is plenty of room for improvement within the algorithm. using the provided threshold numbers we can optimize this pairs trading stratagey.

notes: data was crammed into a single data frame in order to quickly finish this up, a function that can tell which stock is positive and which is negative could be and additional task to automate the task of pairs trading much more.